+++ cli/commands/root.go
@@ -66,6 +66,7 @@ func init() {
 		NewMetadataCmd(ec),
 		NewMigrateCmd(ec),
 		NewSeedCmd(ec),
+		NewDeployCmd(ec),
 		NewActionsCmd(ec),
 		NewPluginsCmd(ec),
 		NewVersionCmd(ec),