+++ server/tests-py/test_metadata.py
@@ -194,14 +194,14 @@ class TestMetadata:
 
         check_query_f(hge_ctx, self.dir() + '/pg_track_function_with_comment_teardown.yaml')
 
-    def test_validate_webhook_transform_success(self, hge_ctx):
-        check_query_f(hge_ctx, self.dir() + '/validate_webhook_transform_success.yaml')
+    def test_test_webhook_transform_success(self, hge_ctx):
+        check_query_f(hge_ctx, self.dir() + '/test_webhook_transform_success.yaml')
 
-    def test_validate_webhook_transform_bad_parse(self, hge_ctx):
-        check_query_f(hge_ctx, self.dir() + '/validate_webhook_transform_bad_parse.yaml')
+    def test_test_webhook_transform_bad_parse(self, hge_ctx):
+        check_query_f(hge_ctx, self.dir() + '/test_webhook_transform_bad_parse.yaml')
 
-    def test_validate_webhook_transform_bad_eval(self, hge_ctx):
-        check_query_f(hge_ctx, self.dir() + '/validate_webhook_transform_bad_eval.yaml')
+    def test_test_webhook_transform_bad_eval(self, hge_ctx):
+        check_query_f(hge_ctx, self.dir() + '/test_webhook_transform_bad_eval.yaml')
 
     @pytest.mark.skipif(
         os.getenv('HASURA_GRAPHQL_PG_SOURCE_URL_1') == os.getenv('HASURA_GRAPHQL_PG_SOURCE_URL_2') or